chore: polish `/tools` UX
