+++ platform/frontend/src/app/tools/layout.tsx
@@ -11,10 +11,6 @@ export default function ToolsLayout({
     <PageLayout
       title="Tools"
       description="Tools displayed here are either detected from requests between agents and LLMs or sourced from installed MCP servers."
-      tabs={[
-        { label: "Agents Assigned", href: "/tools/agents-assigned" },
-        { label: "Without Agents", href: "/tools/without-agents" },
-      ]}
     >
       {children}
     </PageLayout>
