--- platform/frontend/src/app/tools/without-agents/page.tsx
@@ -1,7 +1,0 @@
-import { WithoutAgentsClient } from "./page.client";
-
-export const dynamic = "force-dynamic";
-
-export default async function WithoutAgentsPage() {
-  return <WithoutAgentsClient />;
-}
